Improved Chart Setup
