+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "pooltogether-stats",
-	"version": "2.1.4",
+	"version": "2.2.0",
 	"type": "module",
 	"private": true,
 	"scripts": {
